Extract mapping code for Inventory pull response into dedicated class
